get_guess now in client and not in client logic
